Extract get_base_port_from_config get_chain_ports_from_config core.config.utils
