+++ core/checks/schain.py
@@ -17,13 +17,19 @@
 #   You should have received a copy of the GNU Affero General Public License
 #   along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-import os
 import logging
+import os
 import time
 from typing import Any, List, Optional
 
 from skale.types.schain import SchainName
 
+from core.chain.runner import (
+    get_container_name,
+    get_ima_container_time_frame,
+    get_image_name,
+    is_new_image_pulled,
+)
 from core.checks.base import (
     API_ALLOWED_CHECKS,
     BaseSkaledChecks,
@@ -32,28 +38,21 @@ from core.checks.base import (
     log_checks_dict,
     save_checks_dict,
 )
-from core.node import ExtendedManagerNodeInfo, get_current_ips
 from core.config.schain.file_manager import ConfigFileManager
 from core.config.schain.helper import (
-    get_base_port_from_config,
     get_node_ips_from_config,
     get_own_ip_from_config,
 )
-from core.schains.dkg.utils import get_secret_key_share_filepath
+from core.config.utils import get_base_port_from_config
 from core.firewall import IRuleController
-from core.schains.ima import get_ima_time_frame, get_migration_ts as get_ima_migration_ts
+from core.node import ExtendedManagerNodeInfo, get_current_ips
+from core.schains.dkg.utils import get_secret_key_share_filepath
 from core.schains.external_config import ExternalConfig, ExternalState
-from core.chain.runner import (
-    get_container_name,
-    get_ima_container_time_frame,
-    get_image_name,
-    is_new_image_pulled,
-)
-
+from core.schains.ima import get_ima_time_frame
+from core.schains.ima import get_migration_ts as get_ima_migration_ts
 from tools.configs.containers import IMA_CONTAINER
 from tools.docker_utils import DockerUtils
 from tools.resources import get_statsd_client
-
 from web.models.schain import SChainRecord
 
 logger = logging.getLogger(__name__)
